Move provider interface to migrate package
