+++ workflow/v1_2/provider_test.go
@@ -11,11 +11,11 @@ import (
 	"github.com/antha-lang/antha/laboratory/effects"
 	"github.com/antha-lang/antha/logger"
 	"github.com/antha-lang/antha/workflow"
-	"github.com/antha-lang/antha/workflow/migrate/provider"
+	"github.com/antha-lang/antha/workflow/migrate"
 	"github.com/antha-lang/antha/workflow/v1_2"
 )
 
-func getTestProvider() (provider.WorkflowProvider, error) {
+func getTestProvider() (migrate.WorkflowProvider, error) {
 	oldWorkflowPath := filepath.Join("testdata", "sample_v1_2_workflow.json")
 
 	tmpDir, err := ioutil.TempDir("", "tests")
